+++ R/gformula.R
@@ -16,6 +16,8 @@
 #' Users need to indicate whether to treat competing events as censoring events with the \code{compevent_cens} parameter.
 #' Finally, users can specify how to truncate IP weights with the \code{ipw_cutoff_quantile} or \code{ipw_cutoff_value} parameters.
 #'
+#' In addition to the package output described in McGrath et al. (2020), the output will display estimates of the "cumulative percent intervened on" and the "average percent intervened on". When using a custom intervention function, users need to specify whether each individual at that time point is eligible to contribute person-time to the percent intervened on calculations. Specifically, this must be specified in the \code{eligible_pt} column of \code{newdf}. By default, \code{eligible_pt} is set to \code{TRUE} for each individual at each time point in custom interventions.
+#'
 #' @param id                      Character string specifying the name of the ID variable in \code{obs_data}.
 #' @param time_points             Number of time points to simulate. By default, this argument is set equal to the maximum
 #'                                number of records that \code{obs_data} contains for any individual plus 1.
@@ -122,9 +124,12 @@
 #' @param show_progress           Logical scalar indicating whether to print a progress bar for the number of bootstrap samples completed in the R console. This argument is only applicable when \code{parallel} is set to \code{FALSE} and bootstrap samples are used (i.e., \code{nsamples} is set to a value greater than 0). The default is \code{TRUE}.
 #' @param ipw_cutoff_quantile     Percentile by which to truncate inverse probability weights. The default is \code{NULL} (i.e., no truncation). See "Details".
 #' @param ipw_cutoff_value        Cutoff value by which to truncate inverse probability weights. The default is \code{NULL} (i.e., no truncation). See "Details".
+#' @param int_visit_type          Vector of logicals. The kth element is a logical specifying whether to carry forward the intervened value (rather than the natural value) of the treatment variables(s) when performing a carry forward restriction type for the kth intervention in \code{interventions}.
+#'                                When the kth element is set to \code{FALSE}, the natural value of the treatment variable(s) in the kth intervention in \code{interventions} will be carried forward.
+#'                                By default, this argument is set so that the intervened value of the treatment variable(s) is carried forward for all interventions.
 #' @param ...                     Other arguments, which are passed to the functions in \code{covpredict_custom}.
 #' @return                        An object of class \code{gformula_survival}. The object is a list with the following components:
-#' \item{result}{Results table. For survival outcomes, this contains the estimated risk, risk difference, and risk ratio for all interventions (inculding the natural course) at each time point. For continuous end-of-follow-up outcomes, this contains estimated mean outcome, mean difference, and mean ratio for all interventions (inculding natural course) at the last time point. For binary end-of-follow-up outcomes, this contains the estimated outcome probability, probability difference, and probability ratio for all interventions (inculding natural course) at the last time point. If bootstrapping was used, the results table includes the bootstrap risk / mean / probability difference, ratio, standard error, and 95\% confidence interval.}
+#' \item{result}{Results table. For survival outcomes, this contains the estimated risk, risk difference, and risk ratio for all interventions (inculding the natural course) at each time point. For continuous end-of-follow-up outcomes, this contains estimated mean outcome, mean difference, and mean ratio for all interventions (inculding natural course) at the last time point. For binary end-of-follow-up outcomes, this contains the estimated outcome probability, probability difference, and probability ratio for all interventions (inculding natural course) at the last time point. For all outcome types, this also contains the "cumulative percent intervened on" and the "average percent intervened on". If bootstrapping was used, the results table includes the bootstrap risk / mean / probability difference, ratio, standard error, and 95\% confidence interval.}
 #' \item{coeffs}{A list of the coefficients of the fitted models.}
 #' \item{stderrs}{A list of the standard errors of the coefficients of the fitted models.}
 #' \item{vcovs}{A list of the variance-covariance matrices of the parameters of the fitted models.}
@@ -350,7 +355,7 @@ gformula <- function(obs_data, id, time_points = NULL,
                      nsamples = 0, parallel = FALSE, ncores = NA,
                      ci_method = 'percentile', threads, model_fits = FALSE,
                      boot_diag = FALSE, show_progress = TRUE, ipw_cutoff_quantile = NULL,
-                     ipw_cutoff_value = NULL, ...){
+                     ipw_cutoff_value = NULL, int_visit_type = NULL, ...){
   if (! outcome_type %in% c('survival', 'continuous_eof', 'binary_eof')){
     stop("outcome_type must be 'survival', 'continuous_eof', or 'binary_eof', but outcome_type was set to", outcome_type)
   }
@@ -380,7 +385,7 @@ gformula <- function(obs_data, id, time_points = NULL,
                       ci_method = ci_method, threads = threads,
                       model_fits = model_fits, boot_diag = boot_diag,
                       show_progress = show_progress, ipw_cutoff_quantile = ipw_cutoff_quantile,
-                      ipw_cutoff_value = ipw_cutoff_value, ...)
+                      ipw_cutoff_value = ipw_cutoff_value, int_visit_type = int_visit_type, ...)
   } else if (outcome_type == 'continuous_eof'){
     gformula_continuous_eof(obs_data = obs_data, id = id,
                             time_name = time_name, covnames = covnames,
@@ -405,7 +410,7 @@ gformula <- function(obs_data, id, time_points = NULL,
                             ci_method = ci_method, threads = threads,
                             model_fits = model_fits, boot_diag = boot_diag,
                             show_progress = show_progress, ipw_cutoff_quantile = ipw_cutoff_quantile,
-                            ipw_cutoff_value = ipw_cutoff_value, ...)
+                            ipw_cutoff_value = ipw_cutoff_value, int_visit_type = int_visit_type, ...)
   } else if (outcome_type == 'binary_eof'){
     gformula_binary_eof(obs_data = obs_data, id = id,
                         time_name = time_name, covnames = covnames,
@@ -428,7 +433,7 @@ gformula <- function(obs_data, id, time_points = NULL,
                         ci_method = ci_method, threads = threads,
                         model_fits = model_fits, boot_diag = boot_diag,
                         show_progress = show_progress, ipw_cutoff_quantile = ipw_cutoff_quantile,
-                        ipw_cutoff_value = ipw_cutoff_value, ...)
+                        ipw_cutoff_value = ipw_cutoff_value, int_visit_type = int_visit_type, ...)
   }
 }
 
@@ -449,6 +454,8 @@ gformula <- function(obs_data, id, time_points = NULL,
 #' When competing events are present, users need to include a column in \code{obs_data} with a time-varying indicator of the competing event variable and need to indicate the name of the competing event variable and the corresponding model statement with parameters \code{compevent_name} and \code{compevent_model}, respectively.
 #' Users need to indicate whether to treat competing events as censoring events with the \code{compevent_cens} parameter.
 #' Finally, users can specify how to truncate IP weights with the \code{ipw_cutoff_quantile} or \code{ipw_cutoff_value} parameters.
+#'
+#' In addition to the package output described in McGrath et al. (2020), the output will display estimates of the "cumulative percent intervened on" and the "average percent intervened on". When using a custom intervention function, users need to specify whether each individual at that time point is eligible to contribute person-time to the percent intervened on calculations. Specifically, this must be specified in the \code{eligible_pt} column of \code{newdf}. By default, \code{eligible_pt} is set to \code{TRUE} for each individual at each time point in custom interventions.
 #'
 #' @param id                      Character string specifying the name of the ID variable in \code{obs_data}.
 #' @param time_points             Number of time points to simulate. By default, this argument is set equal to the maximum
@@ -555,9 +562,12 @@ gformula <- function(obs_data, id, time_points = NULL,
 #' @param show_progress           Logical scalar indicating whether to print a progress bar for the number of bootstrap samples completed in the R console. This argument is only applicable when \code{parallel} is set to \code{FALSE} and bootstrap samples are used (i.e., \code{nsamples} is set to a value greater than 0). The default is \code{TRUE}.
 #' @param ipw_cutoff_quantile     Percentile by which to truncate inverse probability weights. The default is \code{NULL} (i.e., no truncation). See "Details".
 #' @param ipw_cutoff_value        Cutoff value by which to truncate inverse probability weights. The default is \code{NULL} (i.e., no truncation). See "Details".
+#' @param int_visit_type          Vector of logicals. The kth element is a logical specifying whether to carry forward the intervened value (rather than the natural value) of the treatment variables(s) when performing a carry forward restriction type for the kth intervention in \code{interventions}.
+#'                                When the kth element is set to \code{FALSE}, the natural value of the treatment variable(s) in the kth intervention in \code{interventions} will be carried forward.
+#'                                By default, this argument is set so that the intervened value of the treatment variable(s) is carried forward for all interventions.
 #' @param ...                     Other arguments, which are passed to the functions in \code{covpredict_custom}.
 #' @return                        An object of class \code{gformula_survival}. The object is a list with the following components:
-#' \item{result}{Results table containing the estimated risk and risk ratio for all interventions (inculding the natural course) at each time point. If bootstrapping was used, the results table includes the bootstrap mean risk ratio, standard error, and 95\% confidence interval.}
+#' \item{result}{Results table containing the estimated risk and risk ratio for all interventions (inculding the natural course) at each time point as well as the "cumulative percent intervened on" and the "average percent intervened on". If bootstrapping was used, the results table includes the bootstrap mean risk ratio, standard error, and 95\% confidence interval.}
 #' \item{coeffs}{A list of the coefficients of the fitted models.}
 #' \item{stderrs}{A list of the standard errors of the coefficients of the fitted models.}
 #' \item{vcovs}{A list of the variance-covariance matrices of the parameters of the fitted models.}
@@ -706,7 +716,7 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
                               ci_method = 'percentile', threads,
                               model_fits = FALSE, boot_diag = FALSE,
                               show_progress = TRUE, ipw_cutoff_quantile = NULL,
-                              ipw_cutoff_value = NULL, ...){
+                              ipw_cutoff_value = NULL, int_visit_type = NULL, ...){
 
   lag_indicator <- lagavg_indicator <- cumavg_indicator <- c()
   lag_indicator <- update_lag_indicator(covparams$covmodels, lag_indicator)
@@ -929,6 +939,12 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
     comb_int_times <- c(list(list(0:(time_points - 1))), int_times)
   }
 
+  if (is.null(int_visit_type)){
+    int_visit_type <- rep(TRUE, length(comb_interventions))
+  } else {
+    int_visit_type <- c(TRUE, int_visit_type)
+  }
+
   # Simulate pooled-over-time datasets containing covariates, outcome, and risk for each
   # subject
   if (parallel){
@@ -951,7 +967,7 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
                                  subseed = subseed, time_points = time_points,
                                  obs_data = obs_data, parallel = parallel, max_visits = max_visits,
                                  baselags = baselags, below_zero_indicator = below_zero_indicator,
-                                 min_time = min_time, show_progress = FALSE, ...)
+                                 min_time = min_time, show_progress = FALSE, int_visit_type = int_visit_type, ...)
     parallel::stopCluster(cl)
   } else {
     pools <- lapply(seq_along(comb_interventions), FUN = function(i){
@@ -971,7 +987,7 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
                subseed = subseed, time_points = time_points,
                obs_data = obs_data, parallel = parallel, max_visits = max_visits,
                baselags = baselags, below_zero_indicator = below_zero_indicator,
-               min_time = min_time, show_progress = FALSE, ...)
+               min_time = min_time, show_progress = FALSE, int_visit_type = int_visit_type[i], ...)
     })
   }
 
@@ -1060,7 +1076,7 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
                                       max_visits = max_visits, hazardratio = hazardratio, intcomp = intcomp,
                                       boot_diag = boot_diag, nsimul = nsimul, baselags = baselags,
                                       below_zero_indicator = below_zero_indicator, min_time = min_time,
-                                      show_progress = FALSE, ...)
+                                      show_progress = FALSE, int_visit_type = int_visit_type, ...)
       parallel::stopCluster(cl)
     } else {
       if (show_progress){
@@ -1086,7 +1102,7 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
                          max_visits = max_visits, hazardratio = hazardratio, intcomp = intcomp,
                          boot_diag = boot_diag, nsimul = nsimul, baselags = baselags,
                          below_zero_indicator = below_zero_indicator, min_time = min_time,
-                         show_progress = show_progress, pb = pb, ...)
+                         show_progress = show_progress, pb = pb, int_visit_type = int_visit_type, ...)
     }
 
     comb_result <- rbindlist(lapply(final_bs, FUN = function(m){
@@ -1358,7 +1374,8 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
 #' Users need to include a column in \code{obs_data} with a time-varying censoring variable.
 #' Users need to indicate the name of the censoring variable and a model statement for the censoring variable with parameters \code{censor_name} and \code{censor_model}, respectively.
 #' Finally, users can specify how to truncate IP weights with the \code{ipw_cutoff_quantile} or \code{ipw_cutoff_value} parameters.
-
+#'
+#' In addition to the package output described in McGrath et al. (2020), the output will display estimates of the "cumulative percent intervened on" and the "average percent intervened on". When using a custom intervention function, users need to specify whether each individual at that time point is eligible to contribute person-time to the percent intervened on calculations. Specifically, this must be specified in the \code{eligible_pt} column of \code{newdf}. By default, \code{eligible_pt} is set to \code{TRUE} for each individual at each time point in custom interventions.
 #'
 #' @param id                      Character string specifying the name of the ID variable in \code{obs_data}.
 #' @param obs_data                Data table containing the observed data.
@@ -1448,10 +1465,13 @@ gformula_survival <- function(obs_data, id, time_points = NULL,
 #' @param show_progress           Logical scalar indicating whether to print a progress bar for the number of bootstrap samples completed in the R console. This argument is only applicable when \code{parallel} is set to \code{FALSE} and bootstrap samples are used (i.e., \code{nsamples} is set to a value greater than 0). The default is \code{TRUE}.
 #' @param ipw_cutoff_quantile     Percentile by which to truncate inverse probability weights. The default is \code{NULL} (i.e., no truncation). See "Details".
 #' @param ipw_cutoff_value        Cutoff value by which to truncate inverse probability weights. The default is \code{NULL} (i.e., no truncation). See "Details".
+#' @param int_visit_type          Vector of logicals. The kth element is a logical specifying whether to carry forward the intervened value (rather than the natural value) of the treatment variables(s) when performing a carry forward restriction type for the kth intervention in \code{interventions}.
+#'                                When the kth element is set to \code{FALSE}, the natural value of the treatment variable(s) in the kth intervention in \code{interventions} will be carried forward.
+#'                                By default, this argument is set so that the intervened value of the treatment variable(s) is carried forward for all interventions.
 #' @param ...                     Other arguments, which are passed to the functions in \code{covpredict_custom}.
 #'
 #' @return                        An object of class \code{gformula_continuous_eof}. The object is a list with the following components:
-#' \item{result}{Results table containing the estimated mean outcome for all interventions (inculding natural course) at the last time point. If bootstrapping was used, the results table includes the bootstrap end-of-follow-up mean ratio, standard error, and 95\% confidence interval.}
+#' \item{result}{Results table containing the estimated mean outcome for all interventions (inculding natural course) at the last time point as well as the "cumulative percent intervened on" and the "average percent intervened on". If bootstrapping was used, the results table includes the bootstrap end-of-follow-up mean ratio, standard error, and 95\% confidence interval.}
 #' \item{coeffs}{A list of the coefficients of the fitted models.}
 #' \item{stderrs}{A list of the standard errors of the coefficients of the fitted models.}
 #' \item{vcovs}{A list of the variance-covariance matrices of the parameters of the fitted models.}
@@ -1529,7 +1549,7 @@ gformula_continuous_eof <- function(obs_data, id,
                                     ci_method = 'percentile', threads,
                                     model_fits = FALSE, boot_diag = FALSE,
                                     show_progress = TRUE, ipw_cutoff_quantile = NULL,
-                                    ipw_cutoff_value = NULL, ...){
+                                    ipw_cutoff_value = NULL, int_visit_type = NULL, ...){
 
   lag_indicator <- lagavg_indicator <- cumavg_indicator <- c()
   lag_indicator <- update_lag_indicator(covparams$covmodels, lag_indicator)
@@ -1730,6 +1750,12 @@ gformula_continuous_eof <- function(obs_data, id,
     comb_int_times <- c(list(list(0:(time_points - 1))), int_times)
   }
 
+  if (is.null(int_visit_type)){
+    int_visit_type <- rep(TRUE, length(comb_interventions))
+  } else {
+    int_visit_type <- c(TRUE, int_visit_type)
+  }
+
   if (parallel){
     cl <- prep_cluster(ncores = ncores, threads = threads,  covtypes = covtypes)
     pools <- parallel::parLapply(cl, seq_along(comb_interventions), simulate,
@@ -1749,7 +1775,7 @@ gformula_continuous_eof <- function(obs_data, id,
                                  subseed = subseed, time_points = time_points,
                                  obs_data = obs_data, parallel = parallel,
                                  baselags = baselags, below_zero_indicator = below_zero_indicator,
-                                 min_time = min_time, show_progress = FALSE, ...)
+                                 min_time = min_time, show_progress = FALSE, int_visit_type = int_visit_type, ...)
     parallel::stopCluster(cl)
   } else {
     pools <- lapply(seq_along(comb_interventions), FUN = function(i){
@@ -1768,7 +1794,7 @@ gformula_continuous_eof <- function(obs_data, id,
                subseed = subseed, time_points = time_points,
                obs_data = obs_data, parallel = parallel,
                baselags = baselags, below_zero_indicator = below_zero_indicator,
-               min_time = min_time, show_progress = FALSE, ...)
+               min_time = min_time, show_progress = FALSE, int_visit_type = int_visit_type[i], ...)
     })
   }
 
@@ -1822,7 +1848,7 @@ gformula_continuous_eof <- function(obs_data, id,
                                       max_visits = max_visits, hazardratio = hazardratio, intcomp = intcomp,
                                       boot_diag = boot_diag, nsimul = nsimul, baselags = baselags,
                                       below_zero_indicator = below_zero_indicator, min_time = min_time,
-                                      show_progress = FALSE, ...)
+                                      show_progress = FALSE, int_visit_type = int_visit_type, ...)
       parallel::stopCluster(cl)
 
     } else {
@@ -1849,7 +1875,7 @@ gformula_continuous_eof <- function(obs_data, id,
                          max_visits = max_visits, hazardratio = hazardratio, intcomp = intcomp,
                          boot_diag = boot_diag, nsimul = nsimul, baselags = baselags,
                          below_zero_indicator = below_zero_indicator, min_time = min_time,
-                         show_progress = show_progress, pb = pb, ...)
+                         show_progress = show_progress, pb = pb, int_visit_type = int_visit_type, ...)
     }
     comb_result <- rbindlist(lapply(final_bs, FUN = function(m){
       as.data.table(t(m$Result))
@@ -2082,6 +2108,8 @@ gformula_continuous_eof <- function(obs_data, id,
 #' Users need to indicate the name of the censoring variable and a model statement for the censoring variable with parameters \code{censor_name} and \code{censor_model}, respectively.
 #' Finally, users can specify how to truncate IP weights with the \code{ipw_cutoff_quantile} or \code{ipw_cutoff_value} parameters.
 #'
+#' In addition to the package output described in McGrath et al. (2020), the output will display estimates of the "cumulative percent intervened on" and the "average percent intervened on". When using a custom intervention function, users need to specify whether each individual at that time point is eligible to contribute person-time to the percent intervened on calculations. Specifically, this must be specified in the \code{eligible_pt} column of \code{newdf}. By default, \code{eligible_pt} is set to \code{TRUE} for each individual at each time point in custom interventions.
+#'
 #' @param id                      Character string specifying the name of the ID variable in \code{obs_data}.
 #' @param obs_data                Data table containing the observed data.
 #' @param seed                    Starting seed for simulations and bootstrapping.
@@ -2170,10 +2198,13 @@ gformula_continuous_eof <- function(obs_data, id,
 #' @param show_progress           Logical scalar indicating whether to print a progress bar for the number of bootstrap samples completed in the R console. This argument is only applicable when \code{parallel} is set to \code{FALSE} and bootstrap samples are used (i.e., \code{nsamples} is set to a value greater than 0). The default is \code{TRUE}.
 #' @param ipw_cutoff_quantile     Percentile by which to truncate inverse probability weights. The default is \code{NULL} (i.e., no truncation). See "Details".
 #' @param ipw_cutoff_value        Cutoff value by which to truncate inverse probability weights. The default is \code{NULL} (i.e., no truncation). See "Details".
+#' @param int_visit_type          Vector of logicals. The kth element is a logical specifying whether to carry forward the intervened value (rather than the natural value) of the treatment variables(s) when performing a carry forward restriction type for the kth intervention in \code{interventions}.
+#'                                When the kth element is set to \code{FALSE}, the natural value of the treatment variable(s) in the kth intervention in \code{interventions} will be carried forward.
+#'                                By default, this argument is set so that the intervened value of the treatment variable(s) is carried forward for all interventions.
 #' @param ...                     Other arguments, which are passed to the functions in \code{covpredict_custom}.
 #'
 #' @return An object of class \code{gformula_binary_eof}. The object is a list with the following components:
-#' \item{result}{Results table containing the estimated outcome probability for all interventions (inculding natural course) at the last time point. If bootstrapping was used, the results table includes the bootstrap end-of-follow-up mean ratio, standard error, and 95\% confidence interval.}
+#' \item{result}{Results table containing the estimated outcome probability for all interventions (inculding natural course) at the last time point as well as the "cumulative percent intervened on" and the "average percent intervened on". If bootstrapping was used, the results table includes the bootstrap end-of-follow-up mean ratio, standard error, and 95\% confidence interval.}
 #' \item{coeffs}{A list of the coefficients of the fitted models.}
 #' \item{stderrs}{A list of the standard errors of the coefficients of the fitted models.}
 #' \item{vcovs}{A list of the variance-covariance matrices of the parameters of the fitted models.}
@@ -2252,7 +2283,7 @@ gformula_binary_eof <- function(obs_data, id,
                                 ci_method = 'percentile', threads,
                                 model_fits = FALSE, boot_diag = FALSE,
                                 show_progress = TRUE, ipw_cutoff_quantile = NULL,
-                                ipw_cutoff_value = NULL, ...){
+                                ipw_cutoff_value = NULL, int_visit_type = NULL, ...){
 
   lag_indicator <- lagavg_indicator <- cumavg_indicator <- c()
   lag_indicator <- update_lag_indicator(covparams$covmodels, lag_indicator)
@@ -2447,6 +2478,12 @@ gformula_binary_eof <- function(obs_data, id,
     comb_int_times <- c(list(list(0:(time_points - 1))), int_times)
   }
 
+  if (is.null(int_visit_type)){
+    int_visit_type <- rep(TRUE, length(comb_interventions))
+  } else {
+    int_visit_type <- c(TRUE, int_visit_type)
+  }
+
   if (parallel){
     cl <- prep_cluster(ncores = ncores, threads = threads , covtypes = covtypes)
     pools <- parallel::parLapply(cl, seq_along(comb_interventions), simulate,
@@ -2466,7 +2503,7 @@ gformula_binary_eof <- function(obs_data, id,
                                  subseed = subseed, time_points = time_points,
                                  obs_data = obs_data, parallel = parallel,
                                  baselags = baselags, below_zero_indicator = below_zero_indicator,
-                                 min_time = min_time, show_progress = FALSE, ...)
+                                 min_time = min_time, show_progress = FALSE, int_visit_type = int_visit_type, ...)
     parallel::stopCluster(cl)
 
   } else {
@@ -2486,7 +2523,7 @@ gformula_binary_eof <- function(obs_data, id,
                subseed = subseed, time_points = time_points,
                obs_data = obs_data, parallel = parallel,
                baselags = baselags, below_zero_indicator = below_zero_indicator,
-               min_time = min_time, show_progress = FALSE, ...)
+               min_time = min_time, show_progress = FALSE, int_visit_type = int_visit_type[i], ...)
     })
   }
 
@@ -2540,7 +2577,7 @@ gformula_binary_eof <- function(obs_data, id,
                                       max_visits = max_visits, hazardratio = hazardratio, intcomp = intcomp,
                                       boot_diag = boot_diag, nsimul = nsimul, baselags = baselags,
                                       below_zero_indicator = below_zero_indicator, min_time = min_time,
-                                      show_progress = FALSE, ...)
+                                      show_progress = FALSE, int_visit_type = int_visit_type, ...)
       parallel::stopCluster(cl)
 
     } else {
@@ -2567,7 +2604,7 @@ gformula_binary_eof <- function(obs_data, id,
                          max_visits = max_visits, hazardratio = hazardratio, intcomp = intcomp,
                          boot_diag = boot_diag, nsimul = nsimul, baselags = baselags,
                          below_zero_indicator = below_zero_indicator, min_time = min_time,
-                         show_progress = show_progress, pb = pb, ...)
+                         show_progress = show_progress, pb = pb, int_visit_type = int_visit_type, ...)
     }
     comb_result <- rbindlist(lapply(final_bs, FUN = function(m){
       as.data.table(t(m$Result))