+++ R/restrictions.R
@@ -14,6 +14,7 @@
 #'                    the user desires to assign to the covariate when it is not modeled).
 #' @param time_name   Character string specifying the name of the time variable in \code{pool} and \code{newdf}.
 #' @param t           Integer specifying the current time index.
+#' @param ...         This argument is not used in this function.
 #' @return No value is returned. The data table \code{newdf} is modified in place.
 #' @examples
 #' ## Estimating the effect of static treatment strategies on risk of a
@@ -61,7 +62,7 @@
 #'
 #' @import data.table
 #' @export
-simple_restriction <- function(newdf, pool, restriction, time_name, t){
+simple_restriction <- function(newdf, pool, restriction, time_name, t, ...){
   classtmp <- class(newdf[!eval(parse(text = restriction[[2]]))][[restriction[[1]]]])
   myclass <- paste('as.', classtmp, sep = "")
   newdf[!eval(parse(text = restriction[[2]])), (restriction[[1]]) :=
@@ -86,6 +87,8 @@ simple_restriction <- function(newdf, pool, restriction, time_name, t){
 #'                    entry some value used by the function (in this case, this entry is not used).
 #' @param time_name   Character string specifying the name of the time variable in \code{pool} and \code{newdf}.
 #' @param t           Integer specifying the current time index.
+#' @param int_visit_type Logical scalar specifying whether to carry forward the intervened value (rather than the natural value) of the treatment variables(s) when performing a carry forward restriction type
+#' @param intvar A vector specifying the name(s) of the variable(s) to be intervened on.
 #' @return No value is returned. The data table \code{newdf} is modified in place.
 #' @examples
 #' ## Estimating the effect of static treatment strategies on risk of a
@@ -133,11 +136,16 @@ simple_restriction <- function(newdf, pool, restriction, time_name, t){
 #'
 #' @import data.table
 #' @export
-carry_forward <- function(newdf, pool, restriction, time_name, t){
+carry_forward <- function(newdf, pool, restriction, time_name, t, int_visit_type, intvar){
   restrict_ids <- newdf[!eval(parse(text = restriction[[2]]))]$id
   # For restricted individuals, carry covariate value over from prior visit
   classtmp <- class(newdf[newdf$id %in% restrict_ids][[restriction[[1]]]])
   myclass <- paste('as.', classtmp, sep = "")
+  if (!int_visit_type & restriction[[1]] %in% intvar){
+    myvar <- paste0(restriction[[1]], '_natural')
+  } else {
+    myvar <- restriction[[1]]
+  }
   newdf[newdf$id %in% restrict_ids, (restriction[[1]]) :=
-          get(myclass)(pool[pool[[time_name]] == (t - 1)][newdf$id %in% restrict_ids][[restriction[[1]]]])]
+          get(myclass)(pool[pool[[time_name]] == (t - 1)][newdf$id %in% restrict_ids][[myvar]])]
 }