+++ NEWS.md
@@ -4,6 +4,9 @@
 
 -   Added the “cumulative percent intervened on” and “average percent
     intervened on” to the output of the `gformula()` function
+-   Added option for users to carry forward the natural value of
+    treatment rather than the intervened value. See the `int_visit_type`
+    argument in the `gformula()` function
 -   Added option for users to access the bootstrap replicates of the
     parametric g-formula estimates. See the `boot_diag` argument in the
     `gformula()` function.